+++ cmd/frontend/internal/batches/resolvers/resolver.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend/graphqlutil"
 	"strconv"
 
 	"github.com/graph-gophers/graphql-go"
@@ -13,6 +14,7 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/rbac"
+	"github.com/sourcegraph/sourcegraph/internal/types"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/enterprise"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
@@ -2021,24 +2023,14 @@ func (r *Resolver) AvailableBulkOperations(ctx context.Context, args *graphqlbac
 		return nil, errors.New("no changesets provided")
 	}
 
-	unmarshalledBatchChangeID, err := unmarshalBatchChangeID(args.BatchChange)
+	batchChangeID, changesetIDs, err := unmarshalBulkOperationBaseArgs(args.BulkOperationBaseArgs)
 	if err != nil {
 		return nil, err
 	}
 
-	changesetIDs := make([]int64, 0, len(args.Changesets))
-	for _, changesetID := range args.Changesets {
-		unmarshalledChangesetID, err := unmarshalChangesetID(changesetID)
-		if err != nil {
-			return nil, err
-		}
-
-		changesetIDs = append(changesetIDs, unmarshalledChangesetID)
-	}
-
 	svc := service.New(r.store)
 	availableBulkOperations, err = svc.GetAvailableBulkOperations(ctx, service.GetAvailableBulkOperationsOpts{
-		BatchChange: unmarshalledBatchChangeID,
+		BatchChange: batchChangeID,
 		Changesets:  changesetIDs,
 	})
 
@@ -2089,6 +2081,52 @@ func (r *Resolver) MaxUnlicensedChangesets(ctx context.Context) int32 {
 		// The license could not be checked.
 		return 0
 	}
+}
+
+func (r *Resolver) GetChangesetsByIDs(ctx context.Context, args *graphqlbackend.GetChangesetsByIDsArgs) (_ graphqlutil.SliceConnectionResolver[graphqlbackend.ChangesetResolver], err error) {
+	tr, ctx := trace.New(ctx, "Resolver.GetChangesetsByIDs",
+		attribute.String("batchChange", string(args.BatchChange)),
+		attribute.Int("changesets.len", len(args.Changesets)))
+	defer tr.EndWithErr(&err)
+
+	if err := enterprise.BatchChangesEnabledForUser(ctx, r.store.DatabaseDB()); err != nil {
+		return nil, err
+	}
+
+	if len(args.Changesets) == 0 {
+		return nil, errors.New("no changesets provided")
+	}
+
+	batchChangeID, changesetIDs, err := unmarshalBulkOperationBaseArgs(args.BulkOperationBaseArgs)
+	if err != nil {
+		return nil, err
+	}
+
+	svc := service.New(r.store)
+	changesets, err := svc.GetChangesetsByIDs(ctx, batchChangeID, changesetIDs)
+	if err != nil {
+		return nil, err
+	}
+
+	// We store a map of repos to avoid duplicate DB queries to fetch a changeset's repository.
+	// This is simply a lazy cache for this query.
+	reposMap := map[api.RepoID]*types.Repo{}
+
+	cs := make([]graphqlbackend.ChangesetResolver, len(changesets))
+	for i, changeset := range changesets {
+		repo, ok := reposMap[changeset.RepoID]
+		if !ok {
+			repo, err = r.store.Repos().Get(ctx, changeset.RepoID)
+			if err != nil {
+				return nil, err
+			}
+			reposMap[repo.ID] = repo
+		}
+
+		cs[i] = NewChangesetResolver(r.store, r.gitserverClient, r.logger, changeset, repo)
+	}
+
+	return graphqlutil.NewSliceConnectionResolver(cs, len(cs), len(cs)), nil
 }
 
 func parseBatchChangeStates(ss *[]string) ([]btypes.BatchChangeState, error) {